+++ src/tests/blueprint/t_blueprint_mpi_mesh_relay.cpp
@@ -732,7 +732,7 @@ TEST(blueprint_mpi_relay, test_sparse_domains_case_1)
     bool hdf5_enabled = io_protos["io/protocols/hdf5"].as_string() == "enabled";
     if(!hdf5_enabled)
     {
-        CONDUIT_INFO("HDF5 disabled, skipping save_read_mesh_truncate test");
+        CONDUIT_INFO("HDF5 disabled, skipping test_sparse_domains_case_1 test");
         return;
     }
 
@@ -811,7 +811,7 @@ TEST(blueprint_mpi_relay, test_sparse_domains_case_2)
     bool hdf5_enabled = io_protos["io/protocols/hdf5"].as_string() == "enabled";
     if(!hdf5_enabled)
     {
-        CONDUIT_INFO("HDF5 disabled, skipping save_read_mesh_truncate test");
+        CONDUIT_INFO("HDF5 disabled, skipping test_sparse_domains_case_2 test");
         return;
     }
 
@@ -885,6 +885,103 @@ TEST(blueprint_mpi_relay, test_sparse_domains_case_2)
 
     Node opts; // empty for now
     std::string tout_base = "tout_relay_mpi_sparse_case_2_hdf5";
+
+    remove_path_if_exists(tout_base + ".cycle_000000.root");
+    conduit::relay::mpi::io::blueprint::save_mesh(data,
+                                                  tout_base,
+                                                  "hdf5",
+                                                  opts,
+                                                  comm);
+    EXPECT_TRUE(conduit::utils::is_file(tout_base + ".cycle_000000.root"));
+
+
+
+}
+
+//-----------------------------------------------------------------------------
+// note: sparse topo tests are from ascent usecases
+TEST(blueprint_mpi_relay, test_sparse_domains_case_3)
+{
+    Node io_protos;
+    relay::io::about(io_protos["io"]);
+    bool hdf5_enabled = io_protos["io/protocols/hdf5"].as_string() == "enabled";
+    if(!hdf5_enabled)
+    {
+        CONDUIT_INFO("HDF5 disabled, skipping test_sparse_domains_case_3 test");
+        return;
+    }
+
+    MPI_Comm comm = MPI_COMM_WORLD;
+    int par_rank = mpi::rank(comm);
+    int par_size = mpi::size(comm);
+
+    //
+    // Create an example mesh.
+    //
+
+    Node data;
+    ostringstream oss;
+
+    // rank 1 have 3 domains, rank zero none
+    if(par_rank > 0)
+    {
+        // three domains with different topos
+        for(index_t d =0; d<3; d++)
+        {
+            Node &mesh = data.append();
+
+            mesh["state/cycle"] = 0;
+
+            oss.str("");
+            oss << "my_coords_rank_" <<  par_rank << "_" << d;
+            std::string c_name = oss.str();
+
+            oss.str("");
+            oss << "my_topo_rank_" <<  par_rank << "_" << d;
+            std::string t_name = oss.str();
+
+            oss.str("");
+            oss << "my_field_rank_" <<  par_rank << "_" << d;
+            std::string f_name = oss.str();
+
+            // create the coordinate set
+            mesh["coordsets"][c_name]["type"] = "uniform";
+            mesh["coordsets"][c_name]["dims/i"] = 3;
+            mesh["coordsets"][c_name]["dims/j"] = 3;
+            // add origin and spacing to the coordset (optional)
+            mesh["coordsets"][c_name]["origin/x"] = -10.0;
+            mesh["coordsets"][c_name]["origin/y"] = -10.0;
+            mesh["coordsets"][c_name]["spacing/dx"] = 10.0;
+            mesh["coordsets"][c_name]["spacing/dy"] = 10.0;
+
+            // add the topology
+            // this case is simple b/c it's implicitly derived from the coordinate set
+            mesh["topologies"][t_name]["type"] = "uniform";
+            // reference the coordinate set by name
+            mesh["topologies"][t_name]["coordset"] = c_name;
+
+            // add a simple element-associated field
+            mesh["fields"][f_name]["association"] =  "element";
+            // reference the topology this field is defined on by name
+            mesh["fields"][f_name]["topology"] =  t_name;
+            // set the field values, for this case we have 4 elements
+            mesh["fields"][f_name]["values"].set(DataType::float64(4));
+
+            float64 *ele_vals_ptr = mesh["fields"][f_name]["values"].value();
+
+            for(int i=0;i<4;i++)
+            {
+                ele_vals_ptr[i] = float64(d);
+            }
+        }
+
+        Node verify_info;
+        EXPECT_TRUE(conduit::blueprint::mesh::verify(data,verify_info));
+    }
+
+    Node opts;
+    opts["suffix"] = "cycle";
+    std::string tout_base = "tout_relay_mpi_sparse_case_3_hdf5";
 
     remove_path_if_exists(tout_base + ".cycle_000000.root");
     conduit::relay::mpi::io::blueprint::save_mesh(data,